Migrate money use cases from long to BigDecimal

+++ banking-api/src/main/java/io/dabrowa/whitebox/api/commands/CreateAccountCommand.java
@@ -1,4 +1,6 @@
 package io.dabrowa.whitebox.api.commands;
 
-public record CreateAccountCommand(String number, long initialBalance, long overdraftLimit) {
+import java.math.BigDecimal;
+
+public record CreateAccountCommand(String number, BigDecimal initialBalance, BigDecimal overdraftLimit) {
 }
