+++ banking-api/src/main/java/io/dabrowa/whitebox/api/commands/CreditAccountCommand.java
@@ -2,5 +2,7 @@ package io.dabrowa.whitebox.api.commands;
 
 import org.axonframework.modelling.command.TargetAggregateIdentifier;
 
-public record CreditAccountCommand(@TargetAggregateIdentifier String accountNumber, long creditValue) {
+import java.math.BigDecimal;
+
+public record CreditAccountCommand(@TargetAggregateIdentifier String accountNumber, BigDecimal creditValue) {
 }
