+++ banking-api/src/main/java/io/dabrowa/whitebox/api/commands/DebitAccountCommand.java
@@ -2,5 +2,7 @@ package io.dabrowa.whitebox.api.commands;
 
 import org.axonframework.modelling.command.TargetAggregateIdentifier;
 
-public record DebitAccountCommand(@TargetAggregateIdentifier String accountNumber, long debitValue) {
+import java.math.BigDecimal;
+
+public record DebitAccountCommand(@TargetAggregateIdentifier String accountNumber, BigDecimal debitValue) {
 }
