+++ banking-api/src/main/java/io/dabrowa/whitebox/api/queries/DebitTestQuery.java
@@ -1,6 +1,8 @@
 package io.dabrowa.whitebox.api.queries;
 
-public record DebitTestQuery(String accountNumber, long debitValue) {
+import java.math.BigDecimal;
+
+public record DebitTestQuery(String accountNumber, BigDecimal debitValue) {
     public enum DebitTestResult {
         OPERATION_ALLOWED, OPERATION_FORBIDDEN
     }
