+++ banking-query/src/main/java/io/dabrowa/whitebox/query/repository/BalanceRepository.java
@@ -1,14 +1,15 @@
 package io.dabrowa.whitebox.query.repository;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Optional;
 
 public interface BalanceRepository {
-    Optional<Long> getBalance(String accountNumber);
+    Optional<BigDecimal> getBalance(String accountNumber);
 
-    void debit(String accountNumber, long value);
+    void debit(String accountNumber, BigDecimal value);
 
-    void credit(String accountNumber, long value);
+    void credit(String accountNumber, BigDecimal value);
 
     List<String> accountsWithNegativeBalance();
 }
